(PUP-7098) Updated acceptance test for utf-8 characters in catalog
The original version of this test was modified to set the environment
variable LANG to "en_US.UTF-8" on each call to on() and apply_manifest_on().
There are some VM pooler templates that to not set this by default.

In addition, the utf-8 string is now defined by using the unicode values.