(PUP-7165) Handle hiera.yaml v3 in environment root using --strict
This commit changes the behavior when the lookup logic encounters a
hiera.yaml of version 3 in an environment root. The behavior is now
under control of --strict.

When set to `error`, an error will be raised.
When set to `warning` or `off`, a warning will be logged and the
hiera.yaml file is then ignored by the environment data provider.