(PUP-7171) Fix problem with dotted keys hiera.yaml interpolations
The check if hiera.yaml was stable with respect to changing scope
between lookups, didn't take into account that the interpolated keys
could use dot notation. As a result, dotted keys were used for lookups
in the provided scope. When used with `strict_variables = true`, this
resulted in an error. When set to `false`, it degraded performance.