(PUP-7178) Dispatch hiera custom backend 'hiera' calls back to lookup
When a custom backend evaluates an interpolation expression that contains
a 'hiera' or 'alias' function, the resulting lookup must be dispatched
back to the lookup framework to perform a global lookup.

This commit ensures that the classic Hiera::Interpolation class is
patched so that the two functions are replaced with a Proc that performs
the needed dispatch.

This commit also adds a test that involves the eyaml backend that is
configured from a Hiera version 3 configuration.