(PUP-7142) Autoload Pcore types from TypeSet.
This commit adds a search algorithm to the module loader so that it
is capable of finding types that are embedded in TypeSets.

A `TypeSet` is a type and therefore autoloaded just like any other type.
That the file "types/x.pp" in module 'a' will contain the type `A::X`
still holds true. What is different is that if `A::X` is a `TypeSet`
which in turn contains the type `Z`, then the autoloader will find the
type `A::X::Z` by loading the `A::X` TypeSet and resolve it.

A special TypeSet that represents the module root can be added in the
file "types/init_typeset.pp". This file must define a `TypeSet` that
has an unqualified name that equals the capitalized name of the module
where it is defined.