(PUP-7191) consider non-existent services 'stopped' on AIX
Prior to this commit, attempting to manage a non-existent service would result
in an error even when ensuring the service is stopped, ie:

$ puppet resource service foo ensure=stopped
Error: /Service[foo]: Could not evaluate: Cannot get status of foo, error was:
Execution of '/usr/bin/lssrc -s foo' returned 1: 0513-085 The foo Subsystem is
not on file.
Wrapped exception: Execution of '/usr/bin/lssrc -s foo' returned 1: 0513-085
The foo Subsystem is not on file.
$ echo $?
1

This is in contrast to current and long-standing behavior on RHEL and Ubuntu
systems, where doing something like ensure => stopped on a non-existent service
is a no-op, and from the ral's perspective a non-existent service is considered
to be in the 'stopped' state. That said, this behavior is not necessarily
"standard" though it probably should be. For example, the launchd provider both
prints an error *and* considers the service stopped.

This commit updates the AIX behavior so that a non-existent service is
considered stopped rather than failing with an error and/or failing the puppet
run if ensuring the service stopped. Ral attempts to ensure=running on a
non-existent service still provide an error, but return correct puppet now
instead of just erroring.

Signed-off-by: Moses Mendoza <moses@puppet.com>