(pup-7021) establish central string encoding conversion
This commit creates puppet/util/character_encoding, a single location to house
logic related to the management character encoding conversion logic in puppet.
Prior to this commit, conversion logic was just beginning to fan out and
duplicate throughout puppet - with all the associated risk and maintenance
burden. We will leverage this code in a separate effort adding a
character-encoding-aware wrapper for ruby's Etc module and rewiring our
providers to use that instead.

This commit actually only adds a single public method -
'Puppet::Util::CharacterEncoding::convert_to_utf8'.

This method applies a standard heuristic to attempt to convert a (copy of a)
passed in string to UTF-8. That heuristic is summarized and justified as
follows (much of this is left as code-comments in the commit as well):

Given a supplied string:
- If the string has UTF-8 encoding, return it as-is. We have no need to do
  anything.

otherwise, if the string is not UTF-8 already:
- Check if its bytes represent valid UTF-8 as currently constitued (in its
  non-UTF-8 encoding) and if so, limit ourselves to setting the external
  encoding of the string to UTF-8 rather than actually transcoding it. We do
  this to handle a couple scenarios:

The first scenario is that the string was originally valid UTF-8 but the
current puppet run is not in a UTF-8 environment. In this case, the string will
likely have invalid byte sequences (i.e., string.valid_encoding? == false), and
attempting to transcode will fail with Encoding::InvalidByteSequenceError,
referencing the invalid byte sequence in the original, pre-transcode, string.
We might have gotten here, for example, if puppet is run first in a user
context with UTF-8 encoding (setting the "is" value to UTF-8) and then later
run via cron without UTF-8 specified, resulting in running in POSIX or some
other default encoding on many systems. In this scenario we're effectively
best-guessing this string originated as UTF-8 and only set external encoding to
UTF-8 - transcoding would have failed anyway.

The second scenario (more rare/unlikely, I expect) is that this string does NOT
have invalid byte sequences (string.valid_encoding? == true), but would
transcode cleanly into different, also valid unicode.

Our example case is "\u16A0" - "RUNIC LETTER FEHU FEOH FE"
http://www.fileformat.info/info/unicode/char/16A0/index.htm
0xE1 0x9A 0xA0 / 225 154 160

These bytes are valid in ISO-8859-1 but the character they represent transcodes
(incorrectly, but cleanly) in ruby to *different* characters in UTF-8. That's
not what we want if the user intended the original string as UTF-8.  Since
we're effectively moving towards a universal UTF-8 standard, if the string is
valid UTF-8 as currently constituted, we default to assuming the string
originated in UTF-8 and do not transcode it - we only set external encoding.

Otherwise, if the string is neither currently considered UTF-8 nor would be
valid UTF-8 bytes:
- Check if the string is even transcodable, as best we can. We know at least
  one way to determine it is _not_ transcodable, which is if
  string.valid_encoding == false, which will always result in
  InvalidByteSequenceError. We may augment this check in the future if we have
  better ways. If it is transcodable, go ahead and attempt to transcode (byte
  changing conversion) to UTF-8.

If none of our conditions have been met thus far:
- the string has been confirmed to not be UTF-8, nor would be valid UTF-8 with
  its current bytes, nor can be transcoded to UTF-8. In this scenario, we fail,
  as this string cannot be converted to UTF-8.

Signed-off-by: Moses Mendoza <moses@puppet.com>