(PUP-7021) Add UTF-8-aware Etc wrapper
The Ruby Etc module will return various Struct objects containing strings in
the system's locale, or if they are invalid in the current locale, as
ASCII-8BIT (Binary). This leads to inconsistent and problematic behavior when
Puppet uses the Etc module in its providers, particularly when trying to
concatenate or compare strings which now may have different, and often
incompatible, encodings. The nameservice provider in particular makes heavy use
of Etc, which has resulted in numerous bugs reported regarding UTF-8.

Initial attempts to resolve the issues focused on the locations where string
comparison or concatenation was happening, but it became quickly clear that
there are too many places to do this, and the better solution would likely be
to control the problem where the strings first enter the system - at the Etc
module.

This commit adds a Puppet wrapper around ruby's Etc module that leverages
puppet/util/character_encoding to convert the values returned by Etc to UTF-8.
In order to avoid ever letting the original Etc values get into the system (and
with a side benefit of avoiding extraneous object creation) the Structs are
modified in-place.

Per feedback on https://github.com/puppetlabs/puppet/pull/5634, we do not fail
if a Struct member cannot be converted to UTF-8, as this would cause puppet to
fail if an unmanaged property cannot be converted to UTF-8, even if that
property is explicitly not managed

Instead of failing, keep the original string intact in the Struct - do not
modify in any way

Signed-off-by: Moses Mendoza <moses@puppet.com>