(PUP-7212) Fix two issues with ClassLoader and Modules
Prior to this commit, the ClassLoader would only find classes, not
Modules. This commit fixes that problem. Once that was fixed, another,
perhaps more serious problem was discovered. When searching, the
method `const_get` was used in a way that made traversed modules search
for constants in ancestors which lead to that unnecessary loading being
performed by the autoloader.

Example of when this happened:
A search for Puppet::SyntaxCheckers::Base64 would traverse down to
Puppet::SyntaxCheckers but when calling `const_get` from there, it would
find the ::Base64 module at the very top. Since that isn't a class, the
autoloader would get triggered.

The fix provided in this commit is to pass a second `false` parameter to
`const_get`. It's very likely that this also speeds up the search.