(PUP-7212) Add a Pcore Annotation type and an annotate function
This commit adds the `Annotation` type as an Object Type. The actual
implementation class is a subclass of `Puppet::Pops::Adaptable::Adapter`
and may be extended from the Puppet Language as an `Object` type that
uses `Annotation` as its parent.

The commit also adds the function `annotate` that can be used in several
different ways (see documentation in ruby source for the function).