(PUP-5973) Remove monkey patch of Symbol <=> String
This removes the monkey patch that made it possible to compare strings
with symbols as if symbols are strings. That was problematic for many
reasons (see ticket).

The monkey patch is removed to prevent further accidental issues that
are very hard to track down. Anyone relying on the monkey patched Symbol
have been warned with a deprecation warning in earlier versions.