(PUP-7216) Prevent merge strategy from hiera_xxx to override hiera.yaml
A merge strategy that are passed as a parameter in a call to lookup
should always override a merge strategy defined in hiera.yaml. This
works as intended. However, this mechanism was also used by the hiera_xxx
functions to pass a strategy. The `hiera_array` and `hiera_include` will
pass `unique` and `hiera_hash` will pass `hash`, and when called from a
`hiera_xxx` function, the synthesized parameter must *not* override
a merge strategy defined in hiera.yaml.

This commit adds a `#hiera_xxx_call?` method to the `LookupInvocation`,
ensures that it is set for all hiera_xxx calls, and then ensures that
any merge strategy that is passed down in combination with that setting,
has lower priority than the strategy defined in the `hiera.yaml` file.