(PUP-7224) Add Rake task for generating the AST model from Pcore
Adds a Rake task that, aside from generating the full set of ruby
classes for the AST, also munges those classes to overcome the chicken
and egg problem that occurs when a generated class needs to use the
parser to resolve references. The parser model itself can of course not
contain such references.