(PUP-6134) Explicitly locate gemcmd on windows
Puppet is unable to install ruby gems into anything other than Puppet's ruby on Windows.

In Windows, the puppet ruby bin directory is prepended to PATH (in environment.bat) and provider.rb uses the which() method to find the first matching command in the PATH.

This fix calls which("gem") with a PATH without Puppet's paths, and sets gemcmd to the full-path result.

A comprehensive solution would be to implement an optional 'command' attribute for package providers.