(PUP-7224) Rename factory methods method #== to #eq and #[] to #access
Parts of the code was very hard to read with respect to when things
where being compared for equality or actually produced an equality
expression. Same confusion was present regarding hash lookups and
the production of an access expression.

This commit changes the `#==` method on Factory to #eq and the `#[]`
method to `#access` to avoid this confusion.