(PUP-7232) Suppress deprecation warnings when creating SemVer ranges
The constructor for the old deprecated `SemVer` class takes an optional
argument that, when set to true, will suppress the otherwise logged
deprecation warning. When creating some types of `SemVer` ranges, this
flag was not propagated correctly which resulted in warnings when using
Puppet module tool. This commit ensures that the flag is propagated
for all range constructs.