(PUP-7243) Relax temp directory permissions
The previous version of this test created a temp directory for the test
environment, but the directory is not readable by puppetserver when
running as the puppet user. Due to PUP-7102, puppetserver will fail to
start if any directory in the environmentpath is unreadable. This
version changes the permissions of that directory to 0777.