(PUP-7130) Fix handling of DateTime.httpdate(x).to_time
In Ruby versions prior <2.4.0, the statement
```
x = DateTime.httpdate('Sat, 03 Feb 2001 04:05:06 GMT').to_time.to_s
```
produces a local time like "2001-02-03 05:05:06 +0100". In 2.4.0 the
timezone is retained. This commit ensures that the time used for the
http_metadata checksum is UTC and that the test expects UTC. That works,
regardless of Ruby version.