(PUP-7130) Fix problem with File#expand_path with tilde and ENV['HOME']
Setting `ENV['HOME'] = nil` was used to provoke an exception when using
a tilde character in `expand_path`. In 2.4.0, the expansion works OK
even after assigning nil.

NOTE! The exception can be provoked in an acceptance test if HOME set
to empty before the Ruby process is started:
```
$ export HOME=
$ irb
2.4.0 :001 > File.expand_path('~/tmp')
ArgumentError: non-absolute home
```