(PUP-7252) Stop puppet service with helper
This commit updated the `puppet_mcollective_service_management`
test to use the beaker helper `stop_agent` to set up the precondition
of ensuring that the `puppet` service is not running.

Prior to this commit, attempting to stop the `puppet` service with a
call to `puppet apply` would intermittently fail due to an active
puppet run on the agent. This caused the `puppet apply` attempt to
fail.