(PUP-7204) Add support for mapped paths to hiera.yaml
This commit adds support for a new hierarchy entry named `mapped_paths`.
It must be a three element array of strings where the first string
denotes a scope variable that should point to a collection of strings,
the second string is the variable name that will be mapped to each
element of the collection, and the third string is a template where that
variable can be used in interpolation expressions.

Paths are produced by iterating over the collection, mapping each entry
to the variable name in a local scope, and then resolve the template
using the local scope.