(PUP-7282) Modified test to work with the customer manifest
This test did not work for a very long time. It was re-enabled 2 months ago.
It was not correctly modifed to work but the assert did not catch the error.
The environment needed to have a zpool named 'tstpool' available to it. Also
becuase the LDom we are using does not have an 'ip.if.1' I replaces it with
the name of the available unplumbed interface, 'net1'