(maint) Simplify DefaultValidator initialization
The DefaultValidator class provides default values for construction but
constructed all of the objects itself. This requires some intimate
knowledge of the Puppet::SSL::Configuration class that's not
appropriate; the Configuration class itself is a better place to
determine how to create a default object. This commit refactors the
Configuration object to build default objects and simplifies the
DefaultValidator initialization accordingly.