(PUP-7283) Move #ssl_store from Host to Configuration
The Puppet::SSL::Host class is pretty toxic because it has many methods
that trigger unexpected side effects. Over time it has grown to
encompass a number of only casually related functions, and also includes the
logic to initialize SSL on agents, in a rather ad hoc and abstruse
manner. Finally, the method for generating a Host object for the current
node (`Puppet::SSL::Host.localhost`) is unsafe to call as it will
trigger the agent cert generation process.

Meanwhile, we need to get ahold of an SSL store for certificate
verification in a number of places, and in those places we're carting
along instances of Puppet::SSL::Host and Puppet::SSL::Configuration.
Because the SSL configuration is already being passed around we can move
the SSL store generation to the configuration class. Doing this allows
us to omit the use of the Host class where it might be unsafe to access
and simply use the SSL store where needed.