(PUP-7284) Reenable Hiera eyaml version 3 extension option
The eyaml backend, when configured in a Hiera version 3 config, can make
use of a special 'extension' option. The Hiera 5 function counterpart
(now also used from Hiera 3) did not take this option into account. This
commit ensures that if an extension option has been declared for a
backend in a Hiera version 3 config, that option will be honored.