(PUP-7240) Fix bug when using lookup_options with 'global_only' lookup
This commit fixes a bug that caused an "undefined method `[]` for
nil:NilClass" message when using the special _lookup_options_ key
together with an invocation configured to do "global only" lookups.

Using a custom adapter is the only way to trigger this bug. A suspicion
that it could also be triggered by the deprecated _hiera_xxx_ was found
to be ungrounded. Those methods never checks for lookup options (they
used to go directly to Hiera classic).

There is no test in this commit. The problem is triggered by the tests
provided in the previous commit.