(PUP-7285) Add metric_id http option to http report processor
Add a `metric_id` http option to the http request made by the http report
processor to report urls. This will allow Puppet Server to track the amount of
time it spends making these requests. This option is ignored by Puppet's http
client. The value for this metric id is `[:puppet :report :http]`.