(PUP-7283) disable CRL checking when no CRL has been downloaded
Creating an SSL store through `SSL::Configuration#ssl_store` adds CRL
checking if the global `certificate_revocation` setting is set. However
our current SSL initialization is quite tangled, and trying to fetch the
CRL before we have a signed certificate might fail. The ruby puppet
master implementation required certificate authentication in order to
fetch the CRL; this was fixed in puppetserver 2.3.0 (see SERVER-1056)
but short of probing the master for the version we can't determine the
correct behavior.

This commit adds the ability to bypass this by allowing callers to
disable CRL checking by exposing a `:use_crl` option on `#ssl_store`.
This will allow SSL validators that can't safely download the CRL if a
CRL isn't present to perform peer verification, and can enable CRL
checking if the CRL has been downloaded and install in an out of band
manner.