(PUP-7252) Add helper: wait for agent run to finish
This commit adds a helper method `wait_until_agent_is_not_running_on`
to make sure that a puppet agent run in not in progress on the
specified host. This method is then added to all other helpers here
that call `puppet apply` on a specified host.

This is to guard against the condition where `puppet apply` may be
called on a host that may be currently executing an agent run. If
this condition occurs, the `puppet apply` call will fail due to
the collision. This not typically a condition that the tester is
interested in.