(PUP-7283) Add Validator.best_validator factory method
The DefaultValidator class used to be responsible for upgrading the
verification/authentication internally, but this made for cluttered and
questionably safe code. This commit adds a .best_validator method that
can upgrade the returned validator based on the host SSL state and
delegates the old .default_validator method to .best_validator.