(PUP-7297) Remove bad test for lookup and Hiera 3 merge_behavior
This commit removes a test that verified that lookup made use of
merge_behavior and deep_merge_options defined in hiera.yaml. Althoug
this behavior was added in 5.0, it was found erranious. The behavior
configured in hiera.yaml must only be used by the `hiera_hash` function.
It should never be used by lookup. Consequently, that behavior was
removed as part of fixing PUP-7286.