(PUP-7301) Avoid unnecessary type inference
Since type inference might potentially be very expensive (an instance
can be a very large hash or array), it should only be used unless there
are no cheaper alternatives.

This commit aims to limit the number of type inferences that are
performed during normal execution by:

1. Removing the need for inferring argument types in a call. Instead,
   the callable checks if the argument array is an instance of the
   parameter tuple and if the block is an instance of the block type.
2. Adding some `#instance?` methods where it is relatively easy to
   compute the result without inferring the type.

The rake task `benchmark:hiera_include` shows a very drastic improvement
with this commit in place.