(PUP-7303) Fix faulty hiera test for asserting interpolation changes
The test to assert that Hiera detects changes in interpolations was using
a global variable. This was incorrect since global variables cannot
change once they have been set. This commit changes the test so that
the variable is no longer prefixed with '::' and the change is made
using a nested scope.