(PUP-7301) Use if/elsif/else to determine scalar type
This commit changes the logic that determines if an object is a Scalar
or not so that it uses an if/elsif/else statement instead of iterating
over an array of classes. This approach was chosen after writing
performance tests comparing hash lookup, case/when, if/elsif/else and
array iteration.