(PUP-7301) Avoid redundant recursion in when testing recursive data type
The Pcore type for Hiera data values is self recursive. As such, it will
verify a nested collection by traversing all its contents. The current
code didn't take this into account and would instead do another recursive
check (which in turn was checked by the same recursive type). This gave
a large number of duplicated checks of the same data.