(PUP-7305) Add a #cached_file_data method to the lookup context
This commit adds a new method #cached_file_data(path, &block) to the
`Context` class that is passed to all data provider functions. The
method will cache the contents of a file, or the value returned by
yielding the contents to a block is a block is provided, and serve
that cached value in subsequent calls.

The cache is discarded when the environment goes out of scope.
Individual entries in the cache are discarded if the file that they
reference changes with respect to inode, mtime, or size.