(PUP-7313) Add #each_key delegator to Settings and add #value_sym method
This commit adds a delegators for the #each_pair and #each_key methods
to the `Settings` class.

It also splits the #value method in two. The current that accepts both
`String` and `Symbol` arguments and a new method #value_sym that is more
strict and only accepts `Symbol` arguments. The intent is to then use
the latter in cases where the passes parameters are known to be symbols.