(PUP-7306) Accept newline in literal regular expressions
Prior to this, puppet did not recognize newlines inside a regular
expression (the opening '/' and closing '/') would have to be on the
same line.
That is bad because the language specifications states that Puppet
Regexp is compatible with the implementation of Regexp used in Ruby
since 1.9.3 and it accepts newlines.

Without this fix, some regular expressions generated by for example URI
cannot be understood by puppet.

This fix removes the constraint that no newlines between the opening and
closing '/' are allowed.