(PUP-7310) Convert string interpolations to format strings in lib/puppet/face/help.rb
The gettext tooling does not work correctly with interpolated strings.
Any string marked for extraction into the POT file must therefore be
converted to use Ruby's string format syntax. This updates the extracted
strings in `lib/puppet/face/help.rb`, which was used as a proof of
concept for the gettext tooling.