(PUP-7327) Defer extraction of `ResolvedLocation` location attribute
Before this commit, the location attribute of the `ResolvedLocation`
was extracted too early, resulting in unnecessary file access to
determine if a path existed or not. The `exist?` call was sent to the
`Pathname` instead of the `ResolvedLocation` which caches this info.

This would also have lead to problems when implementing URI based
data provider functions, since in contrast to `Pathname`, the URI does
not respond to `#exist?`