(PUP-7329) This change fixes a regex in describe.rb.
Before this fix a long line without spaces in a puppet type description
causes puppet describe to fail at formatting the description.  This is a
side effect of anchoring the match expression to ^, beginning of string,
when it should be anchored with \A, beginning of line.

For example, with this type, (described by modules/mytype_module/lib/puppet/type/mytype.rb):

Puppet::Type.newtype(:mytype) do
  @doc = ". 12345678901234567890123456789012345678901234567890123456789012345678901234567890123456789012345678901234567890 nick@magpie.puppetlabs.lan
**this part should not repeat!**
"
  newparam(:name) do
    isnamevar
    desc " name parameter for mytype"
  end
end

would give this result before the fix:

[root@syqs5cacqx016cd type]# puppet describe mytype

mytype
======
.
**this part should not
**this part should not
7890123456789012345678901234567890123456789012345678901234567890
nick@magpie.puppetlabs.lan
**this part should not repeat!**

Parameters
----------

- **name**
    name parameter for mytype
[root@syqs5cacqx016cd type]#

After the fix, the result is this :

[root@syqs5cacqx016cd type]# puppet describe mytype

mytype
======
.
1234567890123456789012345678901234567890123456789012345678901234567890123456
7890123456789012345678901234567890 nick@magpie.puppetlabs.lan
**this part should not repeat!**

Parameters
----------

- **name**
    name parameter for mytype
[root@syqs5cacqx016cd type]#