(PUP-1334) overwrite corrupted contents_file in filebucket
If Puppet receives a file to back up in filebucket that has the same checksum
as an existing backed up file, Puppet will detect this, validate that the
files are indeed identical by comparing size and content, and if so use the
existing backup for the new file.

If Puppet detects that the existing backup and the new file are NOT the same,
it will fail with an error "Got passed new contents for sum <shared checksum
value>". Originally, this somewhat opaque error was only expected to only be
raised in the event of an sum collision, which was expected to be very unlikely
and infrequent.

However, in practice, a filebucket contents_file can accidentally vary from its
expected state (and check sum value) a number of ways - for example, if puppet
unexpectedly fails during the backup write but has the file handle open, or if
the contents_file becomes corrupt on disk after write.

Prior to this commit, if this occurred, the same collision error was raised, as
Puppet never validated that the contents_file actually did, in fact, match the
expected check sum value. In this case, the error is both confusing and cannot
be recovered from without manually removing the contents_file (the path to
which is never mentioned).

With this commit, we:

* Add the ability to validate that the contents_file matches the expected checksum
* Since this does a checksum calculation which can be expensive, we only
  perform this check if we have received a new file with a checksum value that
  matches an existing backup but has a different size/contents (the previous
  failure case)
* If the existing backup does indeed have the expected checksum, we
  have more likely found a true sum collision - in this case we still fail, but
  offer a more explicit message to the user with the path to the existing
  backup contents_file (with the bucketdir ommitted for security).
* If the existing backup does not have the expected checksum, we assume it is
  corrupt, issue a warning to the user, and overwrite it with the new backup
  (not failing).

To accomplish this without a mess of nested conditionals, this commit added to
and refactored private methods within Puppet::FileBucket::File to be more
intuitive with their actions/return values.
* #verify_identical_file! is updated to simply return true/false for whether
  the two files are identical, rather than returning nil if identical and
  printing a message and raising an exception if not. Since it no longer raises
  an exception as a code path, it is renamed to #verify_identical_file
* #contents_file_matches_checksum? and #copy_bucket_file_to_contents_file
  are broken out as their own private utility methods to reduce the havoc
  inside #save_to_disk
* debug/warning output and exception raising is centralized in #save_to_disk
  rather than spread out amongst utility methods.
* yard tags are added and corrected for new/existing methods

Signed-off-by: Moses Mendoza <moses@puppet.com>