(PUP-7283) Implement UnauthenticatedValidator
This adds a new SSL validator that can perform SSL peer verification and
hostname identity checking, but will not provide for client
authentication. This is needed in cases where we are bootstrapping the
client SSL state and don't yet have a signed certificate but have a CA
certificate and can authenticate the rest of the connection. Since we
can't necessarily rely on the CRL being present at this point CRL
checking is only enabled when the CRL has been downloaded out of band;
this isn't ideal but will be necessary for now. In the future we can
either do version probing of the puppet server to see if we can download
the CRL, or simply attempt to download the CRL and gracefully accept the
failure if cert authentication is required.