(PUP-7283) delegate DefaultValidator functionality
The functionality implemented in DefaultValidator has been split out to
the pre-existing NoValidator, the UnauthenticatedValidator, and the
CertAuthValidator. This allows us to strip out the duplicated
implementation of the DefaultValidator and call out to generate the
appropriate validators based on the client SSL state. In future versions
of Puppet this validator should be dropped entirely.