(PUP-7336) Fix problem with interpolated global dotted key references
Before this commit, Hiera would consider an interpolation expression
in the _hiera.yaml_ that referenced a variable using a global key with
subkey (a dotted key), to be initally set, although it wasn't. This was
caused by the subkey lookup which transferred the nil value into an emtpy
string. As a result, Hiera would skip that global key since global keys
are known to be immutable (once set, they cannot change). This made
Hiera miss such keys that went from undefined to defined.