(maint) Add missing nil check after interoplating mapped paths.
The interpolate method that resolves from `scope` is called directly
when the new Hiera `mapped_paths` are used. That method will return
`nil` when the lookup fails. The problem was revealed when testing
the fix for PUP-7336 on the 4.10.x branch (a test started failing).
This commit fixes this so that the upcoming merge doesn't result in
failing tests.