(PUP-7259) Remove duplicate to_pson methods
The FormatSupport module was included in Puppet::Util::{Log,Metric,TagSet}
in commits 446f39e73a and cb710b562.

Later, the `to_pson` method was added to FormatSupport in a72520df. This
resulted in duplicate `to_pson` methods defined in both FormatSupport
and Log/Metric/TagSet, even though the implementations were exactly the
same.

This commit removes the overridden and duplicated methods from the
subclasses, so that the FormatSupport implementation is used instead.