(PUP-7340) Fix context_spec failure on OSX
A test that wrote new content with the exact same size to a file and
then checked if the file had changed by comparing inode, mtime, and
size, failed on OSX since the time resolution was too low for the rapid
create, read, change, check cycle and the inode stayed the same. This
commit ensures that the size of the file changes.