(PUP-7340) Add tests to assert cache clear on mtime and inode change
Prior to this commit, the only assertion that was made on what triggered
a cache entry to be cleared, was if the file size changed. This commit
adds two additional tests. One that changes the inode (by first removing
the file and then recreating it with the same size), and one that
sleeps long enough to trigger an mtime change on all platforms.