(PUP-7360) Add a minor_version method to make docs links easier
Sometimes error messages link to versioned docs pages. The docs pages
are only versioned by minor numbers, such as `4.9`. Ergo, it would be
nice to have a method to generate those easily. This adds that method.

This "reimplements" some version parsing because attempts to load the
SemanticPuppet library became too heavy and fragile.