(PUP-7182) Improve error messages from Hiera 5
This commit contains an overhaul of the errors raised by the lookup
framework. Most errors are converted into proper issues and added to
lib/puppet/pops/issues.rb and, as such, allowed to be propagated all
the way to the log while retaining location information.

Many errors originating from the hiera.yaml will now also have line
information. This is accomplished by a combination of regexps and a
fairly naive parser that is aware of strings and comments.

The errors that were reported when a value was found that didn't conform
to the Puppet::LookupValue have been changed to explicitly state the
originating location and, when appropriate, the key of the culprit.

Unit tests were added to the unit/functions/lookup_spec.rb to cover all
the added issues.